// src/Dogger/wwwroot/src/api/openapi/models/RepositoryResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * General
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RepositoryResponse
 */
export interface RepositoryResponse {
    /**
     * 
     * @type {string}
     * @memberof RepositoryResponse
     */
    pullDogId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RepositoryResponse
     */
    handle?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RepositoryResponse
     */
    name?: string | null;
}

export function RepositoryResponseFromJSON(json: any): RepositoryResponse {
    return RepositoryResponseFromJSONTyped(json, false);
}

export function RepositoryResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): RepositoryResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'pullDogId': !exists(json, 'pullDogId') ? undefined : json['pullDogId'],
        'handle': !exists(json, 'handle') ? undefined : json['handle'],
        'name': !exists(json, 'name') ? undefined : json['name'],
    };
}

export function RepositoryResponseToJSON(value?: RepositoryResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'pullDogId': value.pullDogId,
        'handle': value.handle,
        'name': value.name,
    };
}